requirements.txt
requests
simplejson
openai
azure-identity